Move from spectral to spectrum but keep imports in spectral
